requirements.txt
Django==5.2.1
psycopg2==2.9.10
pydantic==2.11.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.5.1
user-agents==2.2.0
celery==5.5.2
geoip2==5.1.0
redis==5.3.0
coverage==7.8.0
resend==2.9.0
boto3==1.38.12
botocore==1.38.12
